requirements.txt
streamlit
pandas
python-dotenv

langchain-core
langchain-community
langchain-huggingface

faiss-cpu
sentence-transformers
